Extract `XiaomiMotionManufacturerCluster` out of motion quirks
